+++ client/src/pages/HomePage.tsx
@@ -1,115 +1,19 @@
-import { CardComponent } from '../components/iu/CardComponent'
 import { CambiarCompany } from '../components/DefineCompany'
-import { CardMetas } from '../components/iu/cardMetas'
 import { useAuth } from '../auth/AuthProvider'
-
-const data = [
-  {
-    id: 1,
-    porcentaje: 10,
-    nombre: 'META DEL DIA',
-    venta: '$56,702,929'
-  },
-  {
-    id: 2,
-    porcentaje: 30,
-    nombre: 'VENTA ACTUAL DEL DIA',
-    venta: '$15,908,201'
-  },
-  {
-    id: 3,
-    porcentaje: 10,
-    nombre: 'META DEL MES',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 4,
-    porcentaje: 40,
-    nombre: 'VENTA ACTUAL DEL MES',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 5,
-    porcentaje: 50,
-    nombre: 'META DEL AÑO',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 6,
-    porcentaje: 10,
-    nombre: 'VENTA ACTUAL DEL AÑO',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 7,
-    porcentaje: 10,
-    nombre: 'META DEL DIA',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 8,
-    porcentaje: 10,
-    nombre: 'VENTA ACTUAL DEL DIA',
-    venta: '$1,200,000,000'
-  },
-  {
-    id: 9,
-    porcentaje: 60,
-    nombre: 'META DEL DIA',
-    venta: '$1,200,000,000'
-  }
-]
-
-const data2 = [
-  {
-    id: 1,
-    name: 'Metas Del Día',
-    venta: '50.000.000',
-    porcentaje: 50
-  },
-  {
-    id: 2,
-    name: 'Ventas Del Día',
-    venta: '23.000.234',
-    porcentaje: 60
-  }
-]
-
-const nombre = 'META DEL DIA'
-const cumplimiento = 'Cumplimiento Servired Venta Total Del Día'
-const porcentaje = 60
+import DahsBoard from '../components/DashBoard'
 
 function HomePage (): JSX.Element {
   const { user } = useAuth()
 
   const empresa = user?.empresa
 
+  console.log(user)
   return (
     <>
-      {empresa === 'Multired y Servired'
-        ? (<CambiarCompany />)
-        : (<>
-          <section className='flex px-10 justify-center py-6'>
-            <CardComponent name={nombre} cumplimiento={cumplimiento} porcentaje={porcentaje} />
-          </section>
-
-          <section className='flex px-12 gap-4 pb-4'>
-            {
-              data2.map(item => {
-                return (<CardComponent key={item.id} cumplimiento={item.venta} porcentaje={item.porcentaje} name={item.name} />)
-              })
-            }
-          </section>
-
-          <section className='grid grid-cols-3 mx-10 gap-2'>
-            {
-              // * Se mapea el array de objetos data y se retorna un componente CardMetas por cada objeto
-              data.map(item => {
-                return (<CardMetas key={item.id} nombre={item.nombre} venta={item.venta} porcentaje={item.porcentaje} />)
-              })
-            }
-          </section>
-        </>)
+      {
+        empresa === 'Multired y Servired'
+          ? (<CambiarCompany />)
+          : (<DahsBoard />)
       }
     </>
   )
